split usd conversion service into separate files
