+++ src/cashier_frontend/src/services/types/usdConversion.service.types.ts
@@ -1,0 +1,4 @@
+export type ConversionRates = {
+    tokenToUsd: number | undefined;
+    usdToToken: number | undefined;
+};
