+++ src/cashier_frontend/src/services/usdConversionService/icExplorerUsdConversion.service.ts
@@ -1,0 +1,85 @@
+import icExplorerAxiosClient from "@/axios/axiosClient";
+import {
+    IcExplorerTokenListResponse,
+    IcExplorerTokenListResponseItem,
+} from "../types/icExplorer.types";
+import { AxiosInstance } from "axios";
+import { IUsdConversionService } from "./IUsdConversionService";
+import { ConversionRates } from "../types/usdConversion.service.types";
+
+export class IcExplorerUsdConversionService implements IUsdConversionService {
+    private client: AxiosInstance;
+
+    constructor() {
+        this.client = icExplorerAxiosClient;
+    }
+
+    public async tokenToUsd(
+        principal: string,
+        address: string,
+        amount: number,
+    ): Promise<number | undefined> {
+        const { tokenToUsd: rate } = await this.getConversionRates(principal, address);
+
+        if (rate === undefined) {
+            return undefined;
+        }
+
+        return amount * rate;
+    }
+
+    public async usdToToken(
+        principal: string,
+        address: string,
+        amount: number,
+    ): Promise<number | undefined> {
+        const { usdToToken: rate } = await this.getConversionRates(principal, address);
+
+        if (rate === undefined) {
+            return undefined;
+        }
+
+        return amount * rate;
+    }
+
+    public async getConversionRates(principal: string, address: string): Promise<ConversionRates> {
+        const token = await this._getToken(principal, address);
+
+        if (!token) {
+            return {
+                tokenToUsd: undefined,
+                usdToToken: undefined,
+            };
+        }
+
+        const tokenAmount = parseFloat(token.amount);
+        const tokenUsdAmount = parseFloat(token.valueUSD);
+        const tokenToUsd = tokenUsdAmount / tokenAmount;
+        const usdToToken = 1 / tokenToUsd;
+
+        return {
+            tokenToUsd,
+            usdToToken,
+        };
+    }
+
+    public async _getTokens(principal: string): Promise<IcExplorerTokenListResponseItem[]> {
+        const response = await this.client.post<IcExplorerTokenListResponse>("/holder/user", {
+            page: 1,
+            size: 100,
+            isDesc: false,
+            principal,
+        });
+
+        return response.data.list;
+    }
+
+    public async _getToken(
+        principal: string,
+        address: string,
+    ): Promise<IcExplorerTokenListResponseItem | undefined> {
+        const tokens = await this._getTokens(principal);
+        const token = tokens.find((token) => token.ledgerId === address);
+        return token;
+    }
+}
