+++ src/cashier_frontend/src/services/usdConversionService/index.ts
@@ -1,0 +1,8 @@
+import { DevIcExplorerUsdConversionService } from "./devIcExplorerUsdConversion.service";
+import { IcExplorerUsdConversionService } from "./icExplorerUsdConversion.service";
+import { IUsdConversionService } from "./IUsdConversionService";
+
+export const UsdConversionService: IUsdConversionService =
+    import.meta.env.MODE === "production"
+        ? new IcExplorerUsdConversionService()
+        : new DevIcExplorerUsdConversionService();
